Guard downgrade with table_exists

+++ lib/galaxy/model/migrations/alembic/versions_gxy/338d0e5deb03_add_cleanup_event_user_assoc_table.py
@@ -41,4 +41,5 @@ def upgrade():
 
 
 def downgrade():
-    drop_table(TABLE_NAME)
+    if table_exists(TABLE_NAME, True):
+        drop_table(TABLE_NAME)
